pyproject.toml
[project]
name = "omero-iscc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "zeroc-ice @ https://github.com/glencoesoftware/zeroc-ice-py-macos-universal2/releases/download/20240325/zeroc_ice-3.6.5-cp312-cp312-macosx_11_0_universal2.whl; sys_platform == 'darwin'",
    "zeroc-ice @ https://github.com/glencoesoftware/zeroc-ice-py-win-x86_64/releases/download/20240325/zeroc_ice-3.6.5-cp312-cp312-win_amd64.whl; sys_platform == 'win32'",
    "zeroc-ice @ https://github.com/glencoesoftware/zeroc-ice-py-linux-x86_64/releases/download/20240202/zeroc_ice-3.6.5-cp312-cp312-manylinux_2_28_x86_64.whl; sys_platform == 'linux'",
    "iscc-sum>=0.1.0",
    "omero-py>=5.21.1",
]

[dependency-groups]
dev = [
    "click>=8.2.1",
    "paramiko>=4.0.0",
    "pyyaml>=6.0.2",
]
